[REFACTOR] Move tasks in additional package folder
